requirements.txt
streamlit>=1.36,<2
torch>=2.8,<2.9
torchvision>=0.23,<0.24
numpy>=1.24,<3
pillow>=9.5,<11
opencv-python-headless>=4.8,<5
scikit-learn>=1.2,<2
matplotlib>=3.7,<4
seaborn>=0.12,<1
pandas>=2,<3
bcrypt>=4,<5
typer>=0.12,<0.13
pydantic>=2.8,<3
pydantic-settings>=2.6,<3
python-dotenv>=1.0,<2
fpdf2>=2.7,<3
# Dev tools (optional)
ruff>=0.6,<0.7
black>=24.0.0,<25
pytest>=8,<9
pytest-cov>=5,<6
